requirements.txt
absl-py==0.12.0
acora==2.2
aiohttp==3.6.2
allennlp==1.1.0
allennlp-models==1.1.0
allennlp-runmodel==0.2.1
altair==4.1.0
apturl==0.5.2
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
backcall==0.2.0
base58==2.1.0
bcrypt==3.1.7
bleach==3.1.5
blinker==1.4
blis==0.4.1
boto3==1.14.45
botocore==1.17.45
Brlapi==0.7.0
cachetools==4.2.2
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
clean-text==0.2.1
cleanco==2.0.1
Click==7.0
colorama==0.4.3
command-not-found==0.3
conllu==4.1
cryptography==2.8
cupshelpers==1.0
cymem==2.0.3
Cython==0.29.21
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.15.2
docx==0.2.4
duplicity==0.8.12.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz
entrypoints==0.3
example==0.1.0
fasteners==0.14.1
filelock==3.0.12
find-job-titles @ file:///home/oem/Documents/orgtitle/find_job_titles.tar.xz
fire==0.3.1
flatbuffers==1.12
fpdf==1.7.2
fpdf2==2.0.5
ftfy==5.8
future==0.18.2
gast==0.4.0
gatenlp==0.8
gitdb==4.0.7
GitPython==3.1.17
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
httplib2==0.14.0
idna==2.8
importlib-metadata==1.7.0
iniconfig==1.0.1
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.10.0
job-titles-finder @ file:///home/oem/Documents/job_titles_finder.tar.xz
joblib==0.16.0
jsonnet==0.16.0
jsonpickle==1.4.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
kazam==1.4.5
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==18.0.1
langdetect==1.0.8
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.2
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==1.1.0
meld==3.20.2
mistune==0.8.4
monotonic==1.5
more-itertools==8.4.0
multidict==4.7.6
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.7
netifaces==0.10.4
nltk==3.5
notebook==6.0.3
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.40
opt-einsum==3.3.0
overrides==3.1.0
packaging==20.4
PAM==0.4.2
pandas==1.1.0
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.7.0
PDF-Layout-Scanner==1.3.3
pdf2docx==0.3.2
pdfminer.six==20200517
pdfplumber @ file:///home/oem/Downloads/pdfplumber-stable.zip
pexpect==4.6.0
phonenumbers==8.12.6
pickleshare==0.7.5
Pillow==6.2.2
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.13.0
ptyprocess==0.6.0
py==1.9.0
py-rouge==1.1
pyahocorasick==1.4.0
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycryptodome==3.9.8
pycups==1.9.73
pydeck==0.6.2
Pygments==2.6.1
PyGObject==3.36.0
PyICU==2.4.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyMuPDF==1.17.4
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.16.0
pyspellchecker==0.5.5
pytesseract==0.3.4
pytest==6.0.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-docx==0.8.10
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
regex==2020.7.14
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.2
SecretStorage==2.3.1
Send2Trash==1.5.0
sentencepiece==0.1.91
simplejson==3.16.0
six==1.15.0
sklearn==0.0
smmap==4.0.0
sortedcontainers==2.2.2
spacy==2.2.4
srsly==1.0.2
streamlit==0.82.0
systemd-python==234
table-ocr @ file:///home/oem/Downloads/image-table-ocr-master.zip
tabula-py==2.1.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.8.3
tesseract==0.1.3
testpath==0.4.4
text-to-image==0.0.5
thinc==7.4.0
threadpoolctl==2.1.0
tika==1.24
tokenizers==0.8.1rc1
toml==0.10.1
toolz==0.11.1
torch==1.6.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
transformers==3.0.2
typing-extensions==3.7.4.3
tzlocal==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unicodecsv==0.14.1
Unidecode==1.1.1
urllib3==1.25.8
usb-creator==0.3.7
validators==0.18.2
wadllib==1.3.3
Wand==0.6.2
wasabi==0.7.1
watchdog==2.1.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
word2number==1.1
wrapt==1.12.1
xkit==0.0.0
yarl==1.5.1
zipp==3.1.0
